# Use a regex rather than Babel for the transformation.