--- a/elm-package.json
+++ b/elm-package.json
@@ -1,5 +1,5 @@
 {
-    "version": "2.0.0",
+    "version": "2.0.1",
     "summary": "Use Webpack-powered CSS Modules inside your Elm views",
     "repository": "https://github.com/cultureamp/elm-css-modules-loader.git",
     "license": "BSD3",