--- a/package.json
+++ b/package.json
@@ -29,5 +29,9 @@
   "dependencies": {
     "babel-generator": "^6.18.2",
     "loader-utils": "^1.1.0"
+  },
+  "prettier": {
+    "singleQuote": true,
+    "trailingComma": "es5"
   }
 }