This feels like a step backwards (I'd almost always prefer parsing to regex transforms), but this loader performs a very simple transformation on very large JS output, and is necessary for build performance on large elm projects.

Reduced build time:

- In the example project
     - From ~2000ms
     - To ~900ms
- In Culture Amp's app build, no cache
    - No cache: From 74s to 68s
    - Cache, full build: From 9.4s to 8.9s
    - Dev server rebuild: From 6.3s to 5.5s

Things to consider:

- The regex might not be perfect - is it worth creating a test suite with some likely edge-cases?
- Using a regex will break column-level source maps. In future it might break 
- Even though the Elm API doesn't change, we should be careful to expose this as a new major version in semver so people aren't pushed to a new version without realising. 
- When we release we should pin to Elm 0.18.