@jasononeil 2017-10-18
![](https://github.trello.services/images/mini-trello-icon.png) [Optimise slow webpack custom loaders](https://trello.com/c/SPQ8jhzv/160-optimise-slow-webpack-custom-loaders)